HitEffect描画

+++ source/Characters/Player.hpp
@@ -73,7 +73,7 @@ private:
 
 	double alpha;
 	double movement_speed;
-	float attack_power;
+	double attack_power;
 
 	SpriteAnimator spriteAnimator;
 
@@ -107,7 +107,7 @@ public:
 	// バフステータス
 	void SetPlayerSpeed(double speed) { movement_speed += speed; };
 	void SetPlayerDamageUp(double damage_up) { attack_power += damage_up; };
-	float GetPlayerAttackDamage() { return attack_power; };
+	double GetPlayerAttackDamage() { return attack_power; };
 
 	// 回避クールタイムの残り秒数を取得
 	double getAvoidanceCooldown() const { return avoidanceCooldown; }
